Rename libraries and sources to be more descriptive
