+++ pyproject.toml
@@ -51,7 +51,7 @@ all = ["python-graphblas", "numpy==2.2.5", "uvloop", "nicegui[plotly]==2.14.1"]
 
 [tool.scikit-build]
 wheel.packages = ["scaler"]
-build.targets=["server", "ymq", "object_storage_server"]
+build.targets=["cc_object_storage_server", "ymq", "object_storage_server"]
 
 [tool.mypy]
 no_strict_optional = true
