+++ scaler/object_storage/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_library(server SHARED)
+add_library(cc_object_storage_server SHARED)
 
-target_sources(server PRIVATE
-  server.cpp
+target_sources(cc_object_storage_server PRIVATE
+  object_storage_server.cpp
   io_helper.cpp
 )
 
@@ -14,11 +14,11 @@ file(MAKE_DIRECTORY "${CAPNPC_OUTPUT_DIR}")
 
 capnp_generate_cpp(objectStorageSources objectStorageHeaders "../protocol/capnp/object_storage.capnp")
 
-target_sources(server PRIVATE ${objectStorageSources})
-target_link_libraries(server PRIVATE CapnProto::capnp)
-target_include_directories(server PRIVATE ${CMAKE_BINARY_DIR})
+target_sources(cc_object_storage_server PRIVATE ${objectStorageSources})
+target_link_libraries(cc_object_storage_server PRIVATE CapnProto::capnp)
+target_include_directories(cc_object_storage_server PRIVATE ${CMAKE_BINARY_DIR})
 
-install(TARGETS server
+install(TARGETS cc_object_storage_server
         LIBRARY DESTINATION ${CMAKE_SOURCE_DIR}/scaler/lib)
 
 find_package(Python COMPONENTS Development)
@@ -29,7 +29,7 @@ set_target_properties(object_storage_server PROPERTIES  PREFIX "")
 
 target_sources(object_storage_server PRIVATE pymod_object_storage_server.cpp)
 target_include_directories(object_storage_server PRIVATE ${Python_INCLUDE_DIRS})
-target_link_libraries(object_storage_server PRIVATE server
+target_link_libraries(object_storage_server PRIVATE cc_object_storage_server
                                             PRIVATE ${Python_LIBRARIES}
 )
