+++ scaler/object_storage/server_interface.h
